asset_backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
